+++ CHANGES.md
@@ -27,6 +27,7 @@
 ##### Additions :tada:
 
 - `Model` can now classify other assets with a given `classificationType`. [#10623](https://github.com/CesiumGS/cesium/pull/10623)
+- `Model` now supports back face culling for point clouds. [#10703](https://github.com/CesiumGS/cesium/pull/10703)
 
 ##### Fixes :wrench: